Extract mechanic review candidate collection
